+++ bootstrapping/build.netfx.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
 
-NOINIT=0
+echo $(bash --version 2>&1 | head -n 1)
+
+#CUSTOMPARAM=0
 BUILD_ARGUMENTS=()
 for i in "$@"; do
     case $(echo $1 | awk '{print tolower($0)}') in
-        -noinit) NOINIT=1;;
+        # -custom-param) CUSTOMPARAM=1;;
         *) BUILD_ARGUMENTS+=("$1") ;;
     esac
     shift
@@ -17,34 +19,28 @@ SCRIPT_DIR=$(cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd)
 # CONFIGURATION
 ###########################################################################
 
-NUGET_VERSION="_NUGET_VERSION_"
 SOLUTION_DIRECTORY="$SCRIPT_DIR/_SOLUTION_DIRECTORY_"
 BUILD_PROJECT_FILE="$SCRIPT_DIR/_BUILD_DIRECTORY_NAME_/_BUILD_PROJECT_NAME_.csproj"
 BUILD_EXE_FILE="$SCRIPT_DIR/_BUILD_DIRECTORY_NAME_/bin/Debug/_BUILD_PROJECT_NAME_.exe"
-
 TEMP_DIRECTORY="$SCRIPT_DIR/_ROOT_DIRECTORY_/.tmp"
 
+NUGET_VERSION="_NUGET_VERSION_"
 NUGET_URL="https://dist.nuget.org/win-x86-commandline/$NUGET_VERSION/nuget.exe"
-NUGET_FILE="$TEMP_DIRECTORY/nuget.exe"
-export NUGET_EXE="$NUGET_FILE"
 
 ###########################################################################
-# PREPARE BUILD
+# EXECUTION
 ###########################################################################
 
-if ! ((NOINIT)); then
-  mkdir -p "$TEMP_DIRECTORY"
-
-  if [ ! -f "$NUGET_FILE" ]; then curl -Lsfo "$NUGET_FILE" $NUGET_URL;
-  elif [ $NUGET_VERSION == "latest" ]; then mono "$NUGET_FILE" update -Self; fi
-
-  mono "$NUGET_FILE" restore "$BUILD_PROJECT_FILE" -SolutionDirectory $SOLUTION_DIRECTORY
+export NUGET_EXE="$TEMP_DIRECTORY/nuget.exe"
+if [ ! -f "$NUGET_EXE" ]; then
+    mkdir -p "$TEMP_DIRECTORY"
+    curl -Lsfo "$NUGET_EXE" "$NUGET_URL"
+elif [ "$NUGET_VERSION" == "latest" ]; then
+    mono "$NUGET_EXE" update -Self
 fi
+echo $("$NUGET_EXE" help 2>&1 | head -n 1)
 
+mono "$NUGET_EXE" restore "$BUILD_PROJECT_FILE" -SolutionDirectory "$SOLUTION_DIRECTORY"
 msbuild "$BUILD_PROJECT_FILE"
-
-###########################################################################
-# EXECUTE BUILD
-###########################################################################
 
 mono "$BUILD_EXE_FILE" ${BUILD_ARGUMENTS[@]}